{{ cookiecutter.repo_name }}/requirements.txt
black
coverage
flake8
pre-commit
pytest
PyYAML